Pull up State to ConfigurableRewriteMojo for use in additional Mojo
As added in https://github.com/openrewrite/rewrite-maven-plugin/pull/726
